moving Variable (from transform::relational) to syntax::term
